Remove runtime plugin test loader workaround
