+++ runtime-plugins/remote-router/package.json
@@ -8,6 +8,7 @@
   },
   "scannableFiles": [
     "./host/config/runtimePlugin.js",
+    "./host/config/runtimePlugin.cjs",
     "./host/config/webpack.config.modulefederation.js",
     "./host/src/features/Home/Home.jsx"
   ]
